5-4. 수정/삭제 기능에 대한 보안 처리 및 LocalDateTime 설정(디비포맷)  - 이전 버전의 치명적인 문제 : 내가 쓴 글인지 체크하는 기능이 필요하다. ex) @DeleteMapping("/{id}")  public String delete(@PathVariable Long id){   qusetionDAO.delete(id);   return "redirect:/index";  }  - 로그인한 사용자인지와 일치하는지 체크필요.
 - Tip : Refactor - extract Local Variable
ex) model.addAttribute("question", qusetionDAO.findOne(id));
 - > 적용시킬 경우 지역 변수로 생성 및 파라미터 자동 변경
QuestionDTO question = qusetionDAO.findOne(id);
  if( question.isSameWriter(sessionedUser)){
   return "/users/loginform";
  }
 model.addAttribute("question", question);

 - 복습 : getter 대신에 메시지 요청으로 확인
 ex) if( question.isSameWriter(sessionedUser))

 - 오.. 매핑해둔 상위 객체자체를 비교함.
  equals() 오버라이딩이 필요하다. why?
  equals : 인스턴스는 다르지만, 값이 같은경우에 true.
  코드 상에서 항상 false로 떨어진다.(모든 인자값에 적용되기 때문에)
public boolean isSameWriter(UserDTO loginUser) {
  return this.writer.equals(loginUser);
 }

 - Source - Generate hashCode() and equals() 로 생성
ex) 조심 : equal 메소드를 호출하는 해당 객체에서 오버라이딩해야지!
아래의 equals() 오버라이딩 이해하기

@Override
 public int hashCode() {
  final int prime = 31;
  int result = 1;
  result = prime * result + ((name == null) ? 0 : name.hashCode());
  return result;
 }

@Override
 public boolean equals(Object obj) {
  if (this == obj)
   return true;
  if (obj == null)
   return false;
  if (getClass() != obj.getClass())
   return false;
  UserDTO other = (UserDTO) obj;
  if (name == null) {
   if (other.name != null)
    return false;
  } else if (!name.equals(other.name))
   return false;
  return true;
 }

 - 아.. 젠장 느낌표 때문에 또 삽질
if ( !questionDTO.isSameWriter(loginUser)) {
 return "/users/loginform";
 }

 - LocalDateTime : 바이너리 타입으로 들어가고 있다. 포맷팅 하는 클래스 필요 -
ex) @Converter(autoApply = true)
public class LocalDateTimeConverter
implements AttributeConverter<LocalDateTime, Timestamp>
 스프링부트가 @Converter 를 참고하여 로컬타임을 스탬프로 매핑함
create_date : 적용 전 : aced000573aad....... / 적욕 후 : 2017-04-20 16:06:50.626
 - import.sql : 테스트용 디폴트 값 설정함
  매핑 관계를 알기 위해 키값을 고정.

+++ src/main/java/net/board/web/QuestionController.java
@@ -19,54 +19,85 @@ import net.board.domain.UserDTO;
 @Controller
 @RequestMapping("/questions")
 public class QuestionController {
-	
+
 	@Autowired
 	QuestionDAO qusetionDAO;
-	
+
 	@GetMapping("/form")
-	public String form(HttpSession session){
-		if ( !HttpSessionUtils.isLoginUser(session)){
-			return "/users/loginform";
+	public String form(HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
 		}
 		return "/qna/form";
 	}
-	
+
 	@PostMapping("")
-	public String create(String title, String contents, HttpSession session){
-		if ( !HttpSessionUtils.isLoginUser(session)){		// 로그인 한 경우에만 글쓰기 가능
-			return "/users/loginform";
+	public String create(String title, String contents, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) { 
+			return "/users/loginForm";
 		}
-		UserDTO sessionedUser = HttpSessionUtils.getUserFromSession(session);							// 사용자 아이디를 알기 위해 
-		QuestionDTO newQuestion = new QuestionDTO(sessionedUser, title, contents); 					// 사용자 아이디=작성자로 dto 생성 
-		qusetionDAO.save(newQuestion);																			// 디비 저장
+		UserDTO sessionedUser = HttpSessionUtils.getUserFromSession(session); 
+		QuestionDTO newQuestion = new QuestionDTO(sessionedUser, title, contents); 
+		qusetionDAO.save(newQuestion); 
 		return "redirect:/index";
 	}
-	
+
 	@GetMapping("/{id}")
-	public String show(@PathVariable Long id, Model model ){
-		model.addAttribute("question", qusetionDAO.findOne(id));	// 해당 질문을 가져와서 뷰 모델로 전달
+	public String show(@PathVariable Long id, Model model) {
+		model.addAttribute("question", qusetionDAO.findOne(id)); 
+																	
 		return "/qna/show";
 	}
-	
+
 	@GetMapping("/{id}/updateForm")
-	public String updateForm(@PathVariable Long id, Model model){
-		model.addAttribute("question", qusetionDAO.findOne(id));	// 해당 질문을 가져와서 뷰 모델로 전달
+	public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
+		}
+
+		UserDTO loginUser = HttpSessionUtils.getUserFromSession(session);
+		QuestionDTO questionDTO = qusetionDAO.findOne(id);
+		if ( !questionDTO.isSameWriter(loginUser)) {
+			return "/users/loginForm";
+		}
+
+		model.addAttribute("question", questionDTO); 
 		return "/qna/updateForm";
-		
+
 	}
-	
+
 	@PutMapping("/{id}")
-	public String update(@PathVariable Long id, String title, String contents){
-		QuestionDTO questionDTO = qusetionDAO.findOne(id); // 질문 가져오기
-		questionDTO.update(title, contents);			// 질문 수정
-		qusetionDAO.save(questionDTO);				// 디비 저장
-		
+	public String update(@PathVariable Long id, String title, String contents, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
+		}
+
+		UserDTO loginUser = HttpSessionUtils.getUserFromSession(session);
+		QuestionDTO questionDTO = qusetionDAO.findOne(id);
+
+		if (questionDTO.isSameWriter(loginUser)) {
+			return "/users/loginForm";
+		}
+		questionDTO.update(title, contents); 
+		qusetionDAO.save(questionDTO); 
+
 		return String.format("redirect:/questions/%d", id);
 	}
-	
+
 	@DeleteMapping("/{id}")
-	public String delete(@PathVariable Long id){
+	public String delete(@PathVariable Long id, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
+		}
+
+		UserDTO loginUser = HttpSessionUtils.getUserFromSession(session);
+		QuestionDTO questionDTO = qusetionDAO.findOne(id);
+		if (questionDTO.isSameWriter(loginUser)) {
+			return "/users/loginForm";
+		}
+
 		qusetionDAO.delete(id);
 		return "redirect:/index";
 	}
+	
 }
